test: shard real Pi lifecycle matrix
